requirements.txt
ipython
psutil
sounddevice
scipy
redis
tensorflow
tensorflow-io
cherrypy
paho-mqtt